@@ -6,7 +6,7 @@ authors = [
 ]
 description = "mpv IPC socket client based on asyncio"
 license = { text="MIT" }
-requires-python = ">=3.9"
+requires-python = ">=3.11"
 classifiers = [
     "Programming Language :: Python :: 3",
     "License :: OSI Approved :: MIT License",
